pyproject.toml
[project]
name = "gemini-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "google-genai>=1.57.0",
    "numpy<2",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sentence-transformers>=5.2.0",
    "torch<2.3.0",
]
